+++ app.rb
@@ -1,111 +1,7 @@
-class Classroom
-  attr_accessor :label
-  attr_reader :students
-
-  def initialize(label)
-    @label = label
-    @students = []
-  end
-
-  def add_students(student)
-    @students.push(student)
-    student.classroom = self
-  end
-end
-
-class Book
-  attr_accessor :title, :author, :rental
-
-  def initialize(title, author)
-    @title = title
-    @author = author
-    @rental = []
-  end
-
-  def add_rental(person, date)
-    Rental.new(date, self, person)
-  end
-end
-
-class Rental
-  attr_accessor :date, :book, :person
-
-  def initialize(date, book, person)
-    @date = date
-
-    @book = book
-    book.rental << self
-
-    @person = person
-    person.rental << self
-  end
-end
-
-class Nameble
-  def correct_name
-    NotImplementedError
-  end
-end
-
-class Person < Nameble
-  attr_reader :id, :rental, :name, :age
-
-  def initialize(age, name = 'Unknown', permission: true)
-    super()
-    @id = Random.rand(1..1000)
-    @name = name
-    @age = age
-    @parent_permission = permission
-    @rental = []
-  end
-
-  def can_use_services?
-    of_age? || @parent_permission
-  end
-
-  def correct_name
-    @name
-  end
-
-  def add_rental(book, date)
-    Rental.new(date, book, self)
-  end
-
-  private
-
-  def of_age?
-    @age >= 18
-  end
-end
-
-class Student < Person
-  attr_reader :classroom
-
-  def initialize(classroom, age, name = 'Unknown', permission: true)
-    super(age, name, permission: permission)
-    @classroom = classroom
-  end
-
-  def classroom=(classroom)
-    @classroom = classroom
-    classroom.students.push(self) unless classroom.students.include?(self)
-  end
-
-  def play_hooky
-    '¯\(ツ)/¯'
-  end
-end
-
-class Teacher < Person
-  def initialize(specialization, age, name = 'Unknown', permission: true)
-    super(age, name, permission: permission)
-    @specialization = specialization
-  end
-
-  def can_use_services?
-    true
-  end
-end
+require_relative './student'
+require_relative './teacher'
+require_relative './book'
+require_relative './rental'
 
 class App
   def initialize
