--- asociations.rb
@@ -1,42 +1,0 @@
-class Classroom
-  attr_accessor :label
-  attr_reader :students
-
-  def initialize(label)
-    @label = label
-    @students = []
-  end
-
-  def add_students(student)
-    @students.push(student)
-    student.classroom = self
-  end
-end
-
-class Book
-  attr_accessor :title, :author, :rentals
-
-  def initialize(title, author)
-    @title = title
-    @author = author
-    @rental = []
-  end
-
-  def add_rental(person, date)
-    Rental.new(date, self, person)
-  end
-end
-
-class Rental
-  attr_accessor :date, :book, :person
-
-  def initialize(date, book, person)
-    @date = date
-
-    @book = book
-    book.rental.push(self)
-
-    @person = person
-    person.rental.push(self)
-  end
-end
