+++ book.rb
@@ -1,0 +1,15 @@
+require_relative './rental'
+
+class Book
+  attr_accessor :title, :author, :rental
+
+  def initialize(title, author)
+    @title = title
+    @author = author
+    @rental = []
+  end
+
+  def add_rental(person, date)
+    Rental.new(date, self, person)
+  end
+end
