+++ main.rb
@@ -1,5 +1,4 @@
-#!/usr/bin/env ruby
-require_relative './app'
+require_relative './starting'
 
 def main
   puts "Welcome to School Library App!\nPlease choose an option by enterin a number from 1 to 7:
@@ -12,24 +11,6 @@ def main
     puts 'Bye!'
   else
     main
-  end
-end
-
-def starting(action)
-  app = App.new
-  case action
-  when 1
-    app.list_all_books
-  when 2
-    app.list_all_people
-  when 3
-    app.create_a_person
-  when 4
-    app.create_a_book
-  when 5
-    app.create_a_rental
-  when 6
-    app.list_all_rentals
   end
 end
