+++ nameble.rb
@@ -1,0 +1,5 @@
+class Nameble
+  def correct_name
+    NotImplementedError
+  end
+end
