+++ student.rb
@@ -1,3 +1,5 @@
+require_relative './person'
+
 class Student < Person
   attr_reader :classroom
